+++ packages/views/CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## Unreleased
 
+### Internal
+
+-   Declare `wpScript: false` to mark the package as bundled rather than externalized. ([#79945](https://github.com/WordPress/gutenberg/pull/79945))
+
 ## 1.17.0 (2026-07-01)